backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
groq
pymupdf
pydantic